move all location/address related classes in one place
https://github.com/metasfresh/metasfresh/issues/5101
