+++ de.metas.adempiere.adempiere/base/src/main/java-legacy/org/compiere/util/Login.java
@@ -51,9 +51,9 @@ import de.metas.acct.api.AcctSchema;
 import de.metas.acct.api.IAcctSchemaDAO;
 import de.metas.acct.api.IPostingService;
 import de.metas.adempiere.model.I_AD_User;
-import de.metas.adempiere.service.ICountryDAO;
 import de.metas.adempiere.service.IPrinterRoutingBL;
 import de.metas.i18n.Language;
+import de.metas.location.ICountryDAO;
 import de.metas.logging.LogManager;
 import de.metas.util.Check;
 import de.metas.util.Services;
