+++ de.metas.adempiere.adempiere/base/src/main/java/de/metas/location/ILocationBL.java
@@ -1,4 +1,4 @@
-package de.metas.adempiere.service;
+package de.metas.location;
 
 /*
  * #%L
