+++ de.metas.adempiere.adempiere/base/src/test/java/de/metas/location/impl/AddressBuilderTest.java
@@ -1,4 +1,4 @@
-package de.metas.adempiere.service.impl;
+package de.metas.location.impl;
 
 import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.model.InterfaceWrapperHelper;
