+++ de.metas.business/src/main/java/de/metas/pricing/exceptions/ProductNotOnPriceListException.java
@@ -4,8 +4,7 @@ import java.time.LocalDate;
 
 import org.adempiere.exceptions.AdempiereException;
 import de.metas.location.CountryId;
-
-import de.metas.adempiere.service.ICountryDAO;
+import de.metas.location.ICountryDAO;
 import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.service.IBPartnerBL;
 import de.metas.i18n.ITranslatableString;
