+++ de.metas.business/src/main/java/de/metas/pricing/rules/campaign_price/callout/C_Campaign_Price.java
@@ -9,11 +9,12 @@ import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
 import org.adempiere.exceptions.FillMandatoryException;
 import de.metas.location.CountryId;
+import de.metas.location.ICountryDAO;
+
 import org.compiere.model.I_C_Campaign_Price;
 import org.compiere.model.ModelValidator;
 import org.springframework.stereotype.Component;
 
-import de.metas.adempiere.service.ICountryDAO;
 import de.metas.bpartner.BPGroupId;
 import de.metas.bpartner.BPartnerId;
 import de.metas.money.CurrencyId;
