+++ de.metas.business/src/main/java/de/metas/tax/api/TaxNotFoundException.java
@@ -6,6 +6,8 @@ import java.util.Date;
 import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.exceptions.AdempiereException;
 import de.metas.location.CountryId;
+import de.metas.location.ICountryDAO;
+
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.service.IOrgDAO;
 import org.adempiere.service.OrgId;
@@ -14,7 +16,6 @@ import org.compiere.model.MLocation;
 import org.compiere.util.Env;
 import org.compiere.util.TimeUtil;
 
-import de.metas.adempiere.service.ICountryDAO;
 import de.metas.i18n.ITranslatableString;
 import de.metas.i18n.ImmutableTranslatableString;
 import de.metas.i18n.TranslatableStringBuilder;
