+++ de.metas.contracts/src/test/java/de/metas/contracts/impl/AbstractFlatrateTermTest.java
@@ -17,6 +17,7 @@ import org.compiere.model.I_C_BPartner;
 import org.compiere.model.I_C_BPartner_Location;
 import org.compiere.model.I_C_Calendar;
 import org.compiere.model.I_C_Country;
+import org.compiere.model.I_C_CountryArea;
 import org.compiere.model.I_C_Currency;
 import org.compiere.model.I_C_Period;
 import org.compiere.model.I_C_Year;
@@ -28,8 +29,6 @@ import org.junit.BeforeClass;
 
 import de.metas.acct.api.AcctSchemaId;
 import de.metas.adempiere.model.I_AD_User;
-import de.metas.adempiere.model.I_C_CountryArea;
-import de.metas.adempiere.service.ICountryAreaBL;
 import de.metas.contracts.IFlatrateBL;
 import de.metas.contracts.flatrate.interfaces.I_C_DocType;
 import de.metas.contracts.impl.FlatrateTermDataFactory.ProductAndPricingSystem;
@@ -44,6 +43,7 @@ import de.metas.inoutcandidate.api.IShipmentScheduleBL;
 import de.metas.inoutcandidate.api.impl.ShipmentScheduleBL;
 import de.metas.invoicecandidate.api.IInvoiceCandidateHandlerBL;
 import de.metas.invoicecandidate.model.I_C_Invoice_Candidate;
+import de.metas.location.ICountryAreaBL;
 import de.metas.util.Services;
 import lombok.Getter;
 import lombok.NonNull;
