+++ de.metas.fresh/de.metas.fresh.base/src/main/java/de/metas/fresh/setup/process/ClientSetup.java
@@ -26,13 +26,13 @@ import org.compiere.util.TrxRunnable;
 import de.metas.acct.api.AcctSchemaId;
 import de.metas.acct.api.IAcctSchemaDAO;
 import de.metas.adempiere.model.I_AD_User;
-import de.metas.adempiere.service.ILocationBL;
 import de.metas.banking.model.I_C_BP_BankAccount;
 import de.metas.banking.service.IBankingBPBankAccountDAO;
 import de.metas.bpartner.service.IBPartnerBL;
 import de.metas.bpartner.service.IBPartnerDAO;
 import de.metas.bpartner.service.IBPartnerOrgBL;
 import de.metas.cache.interceptor.CacheInterceptor;
+import de.metas.location.ILocationBL;
 import de.metas.pricing.service.IPriceListDAO;
 import de.metas.util.Check;
 import de.metas.util.Services;
