+++ de.metas.shipper.gateway.commons/src/main/java/de/metas/shipper/gateway/commons/DeliveryOrderUtil.java
@@ -3,8 +3,8 @@ package de.metas.shipper.gateway.commons;
 import org.adempiere.util.lang.IPair;
 import org.compiere.model.I_C_Location;
 
-import de.metas.adempiere.service.ICountryDAO;
 import de.metas.location.CountryId;
+import de.metas.location.ICountryDAO;
 import de.metas.shipper.gateway.spi.model.Address;
 import de.metas.shipper.gateway.spi.model.Address.AddressBuilder;
 import de.metas.shipper.gateway.spi.model.CountryCode;
