+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/modelvalidator/SwatValidator.java
@@ -76,7 +76,6 @@ import de.metas.adempiere.callout.C_OrderFastInputTabCallout;
 import de.metas.adempiere.engine.MViewModelValidator;
 import de.metas.adempiere.model.I_C_InvoiceLine;
 import de.metas.adempiere.modelvalidator.AD_User;
-import de.metas.adempiere.modelvalidator.C_CountryArea_Assign;
 import de.metas.adempiere.modelvalidator.Order;
 import de.metas.adempiere.modelvalidator.OrderLine;
 import de.metas.adempiere.modelvalidator.OrgInfo;
@@ -101,6 +100,7 @@ import de.metas.invoice.callout.C_InvoiceLine_TabCallout;
 import de.metas.invoicecandidate.api.IInvoiceCandidateListeners;
 import de.metas.invoicecandidate.spi.impl.AttachmentInvoiceCandidateListener;
 import de.metas.invoicecandidate.spi.impl.OrderAndInOutInvoiceCandidateListener;
+import de.metas.location.interceptor.C_CountryArea_Assign;
 import de.metas.logging.LogManager;
 import de.metas.order.document.counterDoc.C_Order_CounterDocHandler;
 import de.metas.report.ReportStarter;
