+++ de.metas.vertical.pharma/src/main/java/de/metas/impexp/product/IFAProductImportHelper.java
@@ -12,7 +12,7 @@ import org.compiere.model.X_I_Product;
 import org.compiere.util.Env;
 import org.compiere.util.TimeUtil;
 
-import de.metas.adempiere.service.ICountryDAO;
+import de.metas.location.ICountryDAO;
 import de.metas.tax.api.ITaxDAO;
 import de.metas.tax.api.ITaxDAO.TaxCategoryQuery;
 import de.metas.tax.api.ITaxDAO.TaxCategoryQuery.VATType;
